+++ src/tataku/gameplay/modes/osu/osu_info/osu_info.rs
@@ -7,13 +7,13 @@ impl GameModeInfo for OsuGameInfo {
     fn new() -> Self { Self }
     fn display_name(&self) -> &'static str { "Osu" }
 
-    fn calc_acc(&self, score: &Score) -> f64 {
-        let x50  = score.judgments.get("x50").copy_or_default()  as f64;
-        let x100 = score.judgments.get("x100").copy_or_default() as f64;
-        let x300 = score.judgments.get("x300").copy_or_default() as f64;
-        let geki = score.judgments.get("geki").copy_or_default() as f64;
-        let katu = score.judgments.get("katu").copy_or_default() as f64;
-        let miss = score.judgments.get("xmiss").copy_or_default() as f64;
+    fn calc_acc(&self, score: &Score) -> f32 {
+        let x50  = score.judgments.get("x50").copy_or_default()  as f32;
+        let x100 = score.judgments.get("x100").copy_or_default() as f32;
+        let x300 = score.judgments.get("x300").copy_or_default() as f32;
+        let geki = score.judgments.get("geki").copy_or_default() as f32;
+        let katu = score.judgments.get("katu").copy_or_default() as f32;
+        let miss = score.judgments.get("xmiss").copy_or_default() as f32;
     
         (50.0 * x50 + 100.0 * (x100 + katu) + 300.0 * (x300 + geki)) 
         / (300.0 * (miss + x50 + x100 + x300 + katu + geki))
